flatten contacts
