+++ lib/xpm_ruby.rb
@@ -10,14 +10,14 @@ require "active_support/core_ext"
 require "builder"
 
 require "xpm_ruby/client"
-require "xpm_ruby/client/contact"
+require "xpm_ruby/contact"
 require "xpm_ruby/connection"
 require "xpm_ruby/job"
 
 require "xpm_ruby/schema/client/add"
 require "xpm_ruby/schema/client/update"
-require "xpm_ruby/schema/client/contact/add"
-require "xpm_ruby/schema/client/contact/update"
+require "xpm_ruby/schema/contact/add"
+require "xpm_ruby/schema/contact/update"
 require "xpm_ruby/schema/job/add"
 require "xpm_ruby/schema/job/delete"
 require "xpm_ruby/schema/job/state"
