--- lib/xpm_ruby/schema/client/contact/update.rb
@@ -1,20 +1,0 @@
-require_relative "../../../types"
-
-module XpmRuby
-  module Schema
-    module Client
-      module Contact
-        Update = Types::Hash.schema(
-          Name?: Types::String,
-          IsPrimary?: Types::String,
-          Salutation?: Types::String,
-          Addressee?: Types::String,
-          Mobile?: Types::String,
-          Email?: Types::String,
-          Phone?: Types::String,
-          Position?: Types::String
-        ).with_key_transform(&:to_sym)
-      end
-    end
-  end
-end
